+++ src/navigation/action/preselectCharacterOne.action.js
@@ -1,0 +1,9 @@
+export const PRESELECT_CHARACTER_ONE = "preselectCharacterOne";
+
+export default function preselectCharacterOne(categoryIndex, characterIndex) {
+  return {
+    type: PRESELECT_CHARACTER_ONE,
+    categoryIndex,
+    characterIndex
+  };
+}