+++ src/navigation/action/preselectCharacterTwo.action.js
@@ -1,0 +1,9 @@
+export const PRESELECT_CHARACTER_TWO = "preselectCharacterTwo";
+
+export default function preselectCharacterTwo(categoryIndex, characterIndex) {
+  return {
+    type: PRESELECT_CHARACTER_TWO,
+    categoryIndex,
+    characterIndex
+  };
+}