+++ src/side/right/selectingCharacterByPlayerOne.presenter.jsx
@@ -9,6 +9,7 @@ import useNavigation from "../../navigation/useData.hook";
 import useNavigationDispatch from "../../navigation/useDispatch.hook";
 import useSelectCharacterSound from "../../configuration/useSelectCharacterSound.hook";
 import useCancelSound from "../../configuration/useCancelSound.hook";
+import preselectCharacterTwo from "../../navigation/action/preselectCharacterTwo.action";
 import selectCharacterTwo from "../../navigation/action/selectCharacterTwo.action";
 import selectCharacterTwoWithStyleAndColor from "../../navigation/action/selectCharacterTwoWithSyleAndColor.action";
 import unselectCharacterOne from "../../navigation/action/unselectCharacterOne.action";
@@ -38,7 +39,15 @@ export default function SelectingCharacterByPlayerOne() {
   const categoryIndex = useCategoryIndex(categories, input, navigation.characterTwoCategoryIndex);
   const category = categories[categoryIndex];
   const characters = category.characters || [];
-  const characterIndex = useCharacterIndex(characters, input, navigation.characterTwoIndex);
+
+  let defaultCharacterIndex = 0;
+  if (navigation.characterTwoPreselectionByCategory && navigation.characterTwoPreselectionByCategory.has(categoryIndex)) {
+    defaultCharacterIndex = navigation.characterTwoPreselectionByCategory.get(categoryIndex);
+  }
+  const onChange = (newCharacterIndex) => {
+    dispatch(preselectCharacterTwo(categoryIndex, newCharacterIndex));
+  };
+  const characterIndex = useCharacterIndex(characters, input, onChange, defaultCharacterIndex);
   const character = characters[characterIndex];
   const characterName = useCharacterName(character);